requirements.txt
Pillow==6.2.0
PuLP==2.0
numpy==1.17.2
pytest==5.2.1
